added new section in linker for memory pool implemented lower buffer size debug logs switched to using custom uart print functions added script for translated debug logs

+++ Core/Inc/task.h
@@ -23,10 +23,10 @@ typedef struct
     uint32_t priority;
     uint32_t og_priority;
     task_state_t state;
-    uint8_t id;
+    int id;
 } tcb_t;
 
-void create_task(tcb_t **task, task_func_t task_func, void *parameters, uint32_t priority, uint32_t *stack, uint32_t stack_size, uint8_t id);
+void create_task(tcb_t **task, task_func_t task_func, void *parameters, uint32_t priority, uint32_t *stack, uint32_t stack_size, int id);
 void update_task_priority(uint32_t priority);
 
 #endif // RTOS_H
